Tarefa added in elastic search and working

+++ xtarefas/src/main/java/com/basis/campina/xtarefas/domain/dto/AnexoDTO.java
@@ -13,7 +13,7 @@ public class AnexoDTO {
 
     private String fileName;
 
-    private int idTarefa;
+    private Long idTarefa;
 
     private String uuid;
 }
